+++ eras/alonzo/impl/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Version history for `cardano-ledger-alonzo`
 
-## 1.0.1.0
+## 1.1.0.0
 
 * Add `ToJSON` instance for `AlonzoTxOut`, `AlonzoScript` and `Datum`
 * Add `ToJSON` instance for `AlonzoPParams StrictMaybe`