+++ eras/conway/impl/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Version history for `cardano-ledger-conway`
 
+## 1.1.0.0
+
+* Add `ToJSON` instance for:
+  * `ConwayTallyState`
+  * `GovernanceAction`
+  * `GovernanceActionState`
+  * `GovernanceActionIx`
+  * `GovernanceActionId`
+* Add `ToJSONKey` instance for `GovernanceActionId`
+
 ## 1.0.0.0
 
 * First properly versioned release.