+++ eras/mary/impl/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Version history for `cardano-ledger-mary`
 
-## 1.0.1.0
+## 1.1.0.0
 
 * Addition of `ToJSON` instances for `AssetName`, `PolicyID`, `MultiAsset` and `MaryValue`.